Add files via upload
COMPLETE EXCEL PROJECT:

This project addressed various key points in Excel, namely cleaning and standardization of data, creation of pivot tables, and last but not least, designing interactive dashboards using pivot tables.

Data cleaning plays an integral role in giving the customer a suitable and accurate service to its business service requirements. In the first section of the project, I cleaned the data using search find and replace to correct the spellings of mispelled words and removed all duplicates that would make repetition trends in the dashboards. Thus, a worksheet was created to be able to cater to those changes and accommodate additional information to assist in the future; furthermore, using Excel formulas to achieve business requirements.
i.e Age_Category, the column is to serve in place of age, so to ease up the category of the people who accumulated what was sold or being able to get a sum of a certain age group.

Creation of pivot tables; I refer to them as table concrete simplifiers. The columns used on the pivot tables give you more thorough information as to the decision you are required to make lies within the columns and rows.i.e., gives you an allowance to distinguish the differences without using formulas.


dashboard using excel, using pivot tables to design charts, then using those charts to access a slicer that will enable interactive dashboards